Header component to a new file
