+++ src/App.jsx
@@ -1,10 +1,12 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 // import assets
 // import editIcon from "./assets/edit.svg";
 // import homeIcon from "./assets/home.svg";
 import smallRupeeIcon from "./assets/small_rupee.svg";
 import largeRupeeIcon from "./assets/large_rupee.svg";
 import PropTypes from "prop-types";
+// components
+import Header from "./Header";
 
 ScrapItems.propTypes = {
 	items: PropTypes.array.isRequired,
@@ -68,47 +70,6 @@ function App() {
 			<Footer />
 		</>
 	);
-}
-
-function Header() {
-	return (
-		<header className="flex flex-col items-center gap-1 w-full">
-			<h1 className="patrick-hand-sc-regular text-6xl text-yellow-700">
-				Scrapulator
-			</h1>
-			<h2 className="patrick-hand-regular text-lg text-yellow-700 tracking-wider">
-				<SubHeading />
-			</h2>
-		</header>
-	);
-}
-
-function SubHeading() {
-	const subHeadings = [
-		"Scrap calculations made easy",
-		"Scrap value at your fingertips",
-		"Calculate your scrap, calculate your earnings",
-	];
-
-	const [currentSubheadingIndex, setCurrentSubheadingIndex] = useState(0);
-
-	// Function to cycle through subheadings
-	const changeSubheading = () => {
-		setCurrentSubheadingIndex(
-			(prevIndex) => (prevIndex + 1) % subHeadings.length
-		);
-	};
-
-	// Change subheading every 5 seconds
-	useEffect(() => {
-		const intervalId = setInterval(() => {
-			changeSubheading();
-		}, 5000); // Change subheading every 5 seconds
-
-		return () => clearInterval(intervalId);
-	});
-
-	return subHeadings[currentSubheadingIndex];
 }
 
 function ScrapItems({ items, onHandleItemEarnings }) {
